// RoomieMatch-frontend/RoomieMatch-frontend/src/app/models/registration.ts
export interface Registration {
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    role_id: number;
    telephone: string;
    bio: string;
    budget: number;
    occupation: number;
    gender: number;
    birthdate: Date;
    currentCityId: number;
    desiredCityId: number;
}